6.20190309   BT_LOST_CC2540F256_V06_20190309(LOST-BLE) a.低电改高低电平，增加上传app优先级

+++ 调试案子/BT-LOST/Firmware/BT-LOST-CC254X/Projects/ble/common/cc2540/OnBoard.h
@@ -75,7 +75,7 @@ typedef struct
 {
   osal_event_hdr_t hdr;
   uint8             state; // shift
-  uint8             keys;  // keys
+  uint16             keys;  // keys  //ABC
 } keyChange_t;
 
 // Timer clock and power-saving definitions
